requirements.txt
python-telegram-bot==20.7
Pillow==10.3.0
moviepy==1.0.3